pyproject.toml
[project]
name = "testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb",
    "fastapi",
    "langchain",
    "langchain-chroma",
    "langchain-community",
    "langchain-google-genai",
    "langchain-google-vertexai",
    "langchain-openai",
    "langchain-tavily",
    "langgraph",
    "python-dotenv",
]
